+++ spec/rubocop/cop/style/avoid_perl_backrefs_spec.rb
@@ -6,13 +6,16 @@ module Rubocop
   module Cop
     module Style
       describe AvoidPerlBackrefs do
-        let(:ap) { AvoidPerlBackrefs.new }
+        let(:cop) { described_class.new }
 
         it 'registers an offence for $1' do
-          inspect_source(ap, ['puts $1'])
-          expect(ap.offences.size).to eq(1)
-          expect(ap.offences.map(&:message))
-            .to eq(['Prefer the use of MatchData over $1.'])
+          inspect_source(cop, ['puts $1'])
+          expect(cop.offences.size).to eq(1)
+        end
+
+        it 'auto-corrects $1 to Regexp.last_match[1]' do
+          new_source = autocorrect_source(cop, '$1')
+          expect(new_source).to eq('Regexp.last_match[1]')
         end
       end
     end